requirements.txt
# Core Web Framework
flask==3.0.3

# Environment Variables
python-dotenv==1.0.1

# Database & ORM (SQLite or any DB)
SQLAlchemy==2.0.36

# Data Handling
pandas==2.2.2
numpy==1.26.4

# LLM/AI APIs (you can choose one or both depending on provider)
openai==1.102.0    # If using OpenAI API
google-generativeai==0.7.2  # If using Gemini API

# Moderation / Validation
pydantic==2.9.2

# HTTP Requests (for APIs, human escalation integration)
requests==2.32.3

# Optional: For running in production
gunicorn==22.0.0

# Development Utilities
black==24.8.0       # code formatter
pytest==8.3.3       # testing

# Waiter
tenacity
pyarrow==17.0.0
fastparquet==2024.5.0